# swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/client": {
      "get": {
        "tags": [],
        "description": "Endpoint para consulta de cliente.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [],
        "description": "Endpoint para cadastro de cliente.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Dados do cliente",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Cliente numero 1"
                },
                "cpf": {
                  "type": "string",
                  "example": "999.888.777-66"
                },
                "phone": {
                  "type": "string",
                  "example": "(48) 9 8877-2211"
                }
              },
              "required": [
                "name",
                "cpf",
                "phone"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/client/{id}": {
      "get": {
        "tags": [],
        "description": "Endpoint para consulta de cliente por id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do Cliente",
            "example": "61ae6508bc916ae4027da93e"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [],
        "description": "Endpoint para atualização de dados do cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do Cliente",
            "example": "61ae6508bc916ae4027da93e"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Dados do cliente",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Cliente numero 1"
                },
                "cpf": {
                  "type": "string",
                  "example": "999.888.777-66"
                },
                "phone": {
                  "type": "string",
                  "example": "(48) 9 8877-2211"
                }
              },
              "required": [
                "name",
                "cpf",
                "phone"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [],
        "description": "Endpoint para excluir cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do Cliente",
            "example": "61ae6508bc916ae4027da93e"
          },
          {
            "name": "schedule_id",
            "description": "Id do Agendamento",
            "type": "string",
            "required": true,
            "in": "path",
            "example": "61ae6508bc916ae4027da93e"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/client/{id}/schedule": {
      "post": {
        "tags": [],
        "description": "Endpoint para agendar um cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do Cliente",
            "example": "61ae6508bc916ae4027da93e"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Dados do agendamento",
            "schema": {
              "type": "object",
              "properties": {
                "time": {
                  "type": "string",
                  "example": "2021-10-11 20:00"
                },
                "duration": {
                  "type": "number",
                  "example": 1
                },
                "description": {
                  "type": "string",
                  "example": "Agendamento numero 1"
                }
              },
              "required": [
                "time",
                "duration",
                "description"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/client/{id}/schedule/{schedule_id}": {
      "delete": {
        "tags": [],
        "description": "Endpoint para excluir agendamento de um cliente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do Cliente",
            "example": "61ae6508bc916ae4027da93e"
          },
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id do Agendamento",
            "example": "61ae6508bc916ae4027da93e"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {}
}